Remove the idea of Message vs. MessageView etc.
Turns out that, with HTTP,
we cannot have a clean separation
of input -> check -> output,
because already at the input stage,
we may want to report notices
(e.g. when we parse HTTP/1.x framing)
or otherwise "write" on the blackboard
that will later be used for checks
(e.g. in HAR input).