Use (Google's) Brotli instead of (Hyper's) brotlipy
Fixes #8.

I don't quite remember now, but I think I chose brotlipy originally
because it had more pre-built wheels. But as of this writing,
Brotli has more wheels (in particular, for Python 3.7),
builds and works fine under PyPy 3.5, and also doesn't require libffi.
And as explained in issue #8, Brotli is a more convenient dependency
than brotlipy.

Minimum version 1.0.1 was chosen to cover python3-brotli=1.0.3
that is shipped with Ubuntu 18.04 LTS.